refactor(naming): rename source files to new naming convention
Apply the agreed file suffix convention across all pods and shared modules.
All 99 tests pass after this rename. No logic changes — pure structural refactor.

Naming convention applied:
  *.client.component.tsx  — CSR components (Three.js, 'use client', browser APIs)
  *.iso.component.tsx     — Isomorphic components (SSR + hydration, no browser deps)
  *.service.ts            — Service factories / Effect-ts service layers
  *.generator.ts          — Pure TS procedural geometry generators
  *.schema.ts             — Zod schema definitions
  *.shader.vert / .frag   — GLSL vertex and fragment shaders
  *.util.ts               — Shared constants and utility modules
  mod.ts                  — Barrel files (unchanged — structural convention)

Renames:
  cms.schemas.ts              → cms.schema.ts
  shared/theme.ts             → shared/theme.util.ts
  scroll-section.ts           → scroll-section.util.ts
  camera-rig.tsx              → camera-rig.client.component.tsx
  experience.tsx              → experience.client.component.tsx
  post-fx.tsx                 → post-fx.client.component.tsx
  galaxy-generator.ts         → galaxy.generator.ts
  galaxy.vert/.frag           → galaxy.shader.vert/.frag
  galaxy-particles.tsx        → galaxy-particles.client.component.tsx
  galaxy-scene.tsx            → galaxy-scene.client.component.tsx
  galaxy-title.tsx            → galaxy-title.client.component.tsx
  city-generator.ts           → city.generator.ts
  rain.vert/.frag             → rain.shader.vert/.frag
  buildings/car-lights/
    neon-signs/rain/
    city-scene.tsx            → *.client.component.tsx
  hero/about/work/contact
    overlay.tsx               → *.iso.component.tsx
  project-card.tsx            → project-card.iso.component.tsx
  audio-service.ts            → audio.service.ts
  audio-toggle.tsx            → audio-toggle.client.component.tsx

Internal imports, mod.ts barrels, test file import paths, fixtures, and
vitest.config.ts coverage excludes all updated to match.
