+++ app/features/city/buildings.client.component.tsx
