+++ app/features/galaxy/galaxy-title.client.component.tsx
