+++ cms/.npmrc
@@ -1,0 +1,1 @@
+{"onlyBuiltDependencies":["esbuild","sharp"]}
